+++ src/resources/views/create.blade.php
@@ -30,7 +30,7 @@
 		    <div class="box-body">
 		      <!-- load the view from the application if it exists, otherwise load the one in the package -->
 		      @if(view()->exists('vendor.backpack.crud.form_content'))
-		      	@include('vendor.backpack.crud.form_content')
+		      	@include('vendor.backpack.crud.form_content', ['fields' => $crud->getFields('create')])
 		      @else
 		      	@include('crud::form_content', ['fields' => $crud->getFields('create')])
 		      @endif