+++ src/resources/views/fields/assets/css/browse.blade.php
@@ -1,2 +1,0 @@
-<!-- include browse server css -->
-<link href="{{ asset('vendor/backpack/colorbox/example2/colorbox.css') }}" rel="stylesheet" type="text/css" />