+++ src/resources/views/fields/assets/css/select2.blade.php
@@ -1,3 +1,0 @@
-<!-- include select2 css-->
-<link href="{{ asset('vendor/backpack/select2/select2.css') }}" rel="stylesheet" type="text/css" />
-<link href="{{ asset('vendor/backpack/select2/select2-bootstrap-dick.css') }}" rel="stylesheet" type="text/css" />