+++ src/resources/views/fields/assets/css/wysiwyg.blade.php
@@ -1,2 +1,0 @@
-<!-- include summernote css-->
-<link href="{{ asset('vendor/backpack/summernote/summernote.css') }}" rel="stylesheet" type="text/css" />