+++ src/resources/views/fields/assets/js/browse.blade.php
@@ -1,3 +1,0 @@
-<!-- include browse server js -->
-<script src="{{ asset('vendor/backpack/colorbox/jquery.colorbox-min.js') }}"></script>
-<script src="{{ asset('vendor/backpack/elfinder/standalonepopup.js') }}"></script>