+++ src/resources/views/fields/page_or_link.blade.php
@@ -98,3 +98,49 @@
     <div class="clearfix"></div>
 
 </div>
+
+
+{{-- ########################################## --}}
+{{-- Extra CSS and JS for this particular field --}}
+{{-- If a field type is shown multiple times on a form, the CSS and JS will only be loaded once --}}
+@if ($crud->checkIfFieldIsFirstOfItsType($field, $fields))
+
+    {{-- FIELD CSS - will be loaded in the after_styles section --}}
+    @push('crud_fields_styles')
+    @endpush
+
+    {{-- FIELD JS - will be loaded in the after_scripts section --}}
+    @push('crud_fields_scripts')
+        <script>
+            jQuery(document).ready(function($) {
+
+                $("#page_or_link_select").change(function(e) {
+                    $(".page_or_link_value input").attr('disabled', 'disabled');
+                    $(".page_or_link_value select").attr('disabled', 'disabled');
+                    $(".page_or_link_value").removeClass("hidden").addClass("hidden");
+
+
+                    switch($(this).val()) {
+                        case 'external_link':
+                            $("#page_or_link_external_link input").removeAttr('disabled');
+                            $("#page_or_link_external_link").removeClass('hidden');
+                            break;
+
+                        case 'internal_link':
+                            $("#page_or_link_internal_link input").removeAttr('disabled');
+                            $("#page_or_link_internal_link").removeClass('hidden');
+                            break;
+
+                        default: // page_link
+                            $("#page_or_link_page select").removeAttr('disabled');
+                            $("#page_or_link_page").removeClass('hidden');
+                    }
+                });
+
+            });
+        </script>
+    @endpush
+
+@endif
+{{-- End of Extra CSS and JS --}}
+{{-- ########################################## --}}